requirements.txt
boto3==1.24.39
botocore==1.27.39
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
chart-studio==1.1.0
click==8.1.3
cycler==0.11.0
dash==2.6.0
dash-auth==1.4.1
dash-bootstrap-components==1.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-loading-spinners==1.0.0
dash-table==5.0.0
Flask==2.1.3
Flask-Compress==1.12
Flask-SeaSurf==1.1.1
fonttools==4.34.4
idna==3.3
imageio==2.20.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
mglearn==0.1.9
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
plotly==5.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
retrying==1.3.3
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==3.1.0
ua-parser==0.15.0
urllib3==1.26.11
Werkzeug==2.2.1
yellowbrick==1.4
zipp==3.8.1

torch
diffusers
transformers
ftfy